dev_requirements.txt
# sample dev-only packages
flake8
black

# setup.py packages used in pyproject.toml
wheel
python-dotenv

-e .